+++ crates/pico_binutils/src/var_int.rs
@@ -17,6 +17,23 @@ impl VarInt {
     pub fn inner(&self) -> i32 {
         self.0
     }
+
+    pub fn to_bytes(&self) -> Vec<u8> {
+        let mut bytes = Vec::with_capacity(5);
+        let mut value = self.0 as u32;
+        loop {
+            let mut temp = (value & (SEGMENT_BITS as u32)) as u8;
+            value >>= 7;
+            if value != 0 {
+                temp |= CONTINUE_BIT;
+            }
+            bytes.push(temp);
+            if value == 0 {
+                break;
+            }
+        }
+        bytes
+    }
 }
 
 impl From<i32> for VarInt {
@@ -89,18 +106,7 @@ impl ReadBytes for VarInt {
 #[cfg(feature = "binary_writer")]
 impl WriteBytes for VarInt {
     fn write(&self, writer: &mut BinaryWriter) -> Result<(), BinaryWriterError> {
-        let mut value = self.0 as u32;
-        loop {
-            let mut temp = (value & (SEGMENT_BITS as u32)) as u8;
-            value >>= 7;
-            if value != 0 {
-                temp |= CONTINUE_BIT;
-            }
-            writer.write(&temp)?;
-            if value == 0 {
-                break;
-            }
-        }
+        writer.write_bytes(self.to_bytes().as_slice())?;
         Ok(())
     }
 }