+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for player skins
 - Added boss bar (1.9+)
 - Support for server icon
+- Network compression
 
 ### Changed